+++ packages/gatsby/src/commands/develop.ts
@@ -16,7 +16,7 @@ import { formatError } from "graphql"
 
 import webpackConfig from "../utils/webpack.config"
 import bootstrap from "../bootstrap"
-import { store, emitter } from "../redux"
+import { store } from "../redux"
 import { syncStaticDir } from "../utils/get-static-dir"
 import { buildHTML } from "./build-html"
 import { withBasePath } from "../utils/path"
@@ -27,13 +27,13 @@ import chalk from "chalk"
 import address from "address"
 import cors from "cors"
 import telemetry from "gatsby-telemetry"
-import WorkerPool from "../utils/worker/pool"
+import * as WorkerPool from "../utils/worker/pool"
 import http from "http"
 import https from "https"
 
 import bootstrapSchemaHotReloader from "../bootstrap/schema-hot-reloader"
 import bootstrapPageHotReloader from "../bootstrap/page-hot-reloader"
-import developStatic from "./develop-static"
+import { developStatic } from "./develop-static"
 import withResolverContext from "../schema/context"
 import sourceNodes from "../utils/source-nodes"
 import createSchemaCustomization from "../utils/create-schema-customization"
@@ -43,7 +43,7 @@ import { slash } from "gatsby-core-utils"
 import { initTracer } from "../utils/tracer"
 import apiRunnerNode from "../utils/api-runner-node"
 import db from "../db"
-import detectPortInUseAndPrompt from "../utils/detect-port-in-use-and-prompt"
+import { detectPortInUseAndPrompt } from "../utils/detect-port-in-use-and-prompt"
 import onExit from "signal-exit"
 import queryUtil from "../query"
 import queryWatcher from "../query/query-watcher"
@@ -52,30 +52,16 @@ import {
   reportWebpackWarnings,
   structureWebpackErrors,
 } from "../utils/webpack-error-utils"
+import { waitUntilAllJobsComplete } from "../utils/wait-until-jobs-complete"
+import {
+  userPassesFeedbackRequestHeuristic,
+  showFeedbackRequest,
+} from "../utils/feedback"
 
 import { BuildHTMLStage, IProgram } from "./types"
-import { waitUntilAllJobsComplete as waitUntilAllJobsV2Complete } from "../utils/jobs-manager"
 
 // checks if a string is a valid ip
 const REGEX_IP = /^(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])\.){3}(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])$/
-
-const waitUntilAllJobsComplete = (): Promise<void> => {
-  const jobsV1Promise = new Promise(resolve => {
-    const onEndJob = (): void => {
-      if (store.getState().jobs.active.length === 0) {
-        resolve()
-        emitter.off(`END_JOB`, onEndJob)
-      }
-    }
-    emitter.on(`END_JOB`, onEndJob)
-    onEndJob()
-  })
-
-  return Promise.all([
-    jobsV1Promise,
-    waitUntilAllJobsV2Complete(),
-  ]).then(() => {})
-}
 
 // const isInteractive = process.stdout.isTTY
 
@@ -353,6 +339,19 @@ async function startServer(program: IProgram): Promise<IServer> {
 }
 
 module.exports = async (program: IProgram): Promise<void> => {
+  // We want to prompt the feedback request when users quit develop
+  // assuming they pass the heuristic check to know they are a user
+  // we want to request feedback from, and we're not annoying them.
+  process.on(
+    `SIGINT`,
+    async (): Promise<void> => {
+      if (await userPassesFeedbackRequestHeuristic()) {
+        showFeedbackRequest()
+      }
+      process.exit(0)
+    }
+  )
+
   if (process.env.GATSBY_EXPERIMENTAL_PAGE_BUILD_ON_DATA_CHANGES) {
     report.panic(
       `The flag ${chalk.yellow(