+++ packages/gatsby-plugin-emotion/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-emotion@4.1.24...gatsby-plugin-emotion@4.1.25) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-plugin-emotion
+
+## [4.1.24](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-emotion@4.1.23...gatsby-plugin-emotion@4.1.24) (2020-03-06)
+
+**Note:** Version bump only for package gatsby-plugin-emotion
+
 ## [4.1.23](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-emotion@4.1.22...gatsby-plugin-emotion@4.1.23) (2020-03-02)
 
 **Note:** Version bump only for package gatsby-plugin-emotion