+++ packages/gatsby-plugin-typescript/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.2.4...gatsby-plugin-typescript@2.2.5) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.2.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.2.3...gatsby-plugin-typescript@2.2.4) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.2.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.2.2...gatsby-plugin-typescript@2.2.3) (2020-03-13)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.2.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.2.1...gatsby-plugin-typescript@2.2.2) (2020-03-07)
+
+### Bug Fixes
+
+- **gatsby-plugin-typescript:** Broader webpack support ([#22003](https://github.com/gatsbyjs/gatsby/issues/22003)) ([4b93826](https://github.com/gatsbyjs/gatsby/commit/4b93826))
+
+## [2.2.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.2.0...gatsby-plugin-typescript@2.2.1) (2020-03-06)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
 # [2.2.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.1.27...gatsby-plugin-typescript@2.2.0) (2020-03-02)
 
 ### Features