+++ packages/gatsby-theme-blog/CHANGELOG.md
@@ -3,6 +3,74 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.30](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.29...gatsby-theme-blog@1.3.30) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.29](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.28...gatsby-theme-blog@1.3.29) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.28](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.27...gatsby-theme-blog@1.3.28) (2020-03-13)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.26...gatsby-theme-blog@1.3.27) (2020-03-13)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.26](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.25...gatsby-theme-blog@1.3.26) (2020-03-13)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.24...gatsby-theme-blog@1.3.25) (2020-03-12)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.24](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.23...gatsby-theme-blog@1.3.24) (2020-03-12)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.23](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.22...gatsby-theme-blog@1.3.23) (2020-03-11)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.21...gatsby-theme-blog@1.3.22) (2020-03-11)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.20...gatsby-theme-blog@1.3.21) (2020-03-11)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.19...gatsby-theme-blog@1.3.20) (2020-03-10)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.18...gatsby-theme-blog@1.3.19) (2020-03-10)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.17...gatsby-theme-blog@1.3.18) (2020-03-09)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.16...gatsby-theme-blog@1.3.17) (2020-03-09)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.15...gatsby-theme-blog@1.3.16) (2020-03-06)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.14...gatsby-theme-blog@1.3.15) (2020-03-06)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.3.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.13...gatsby-theme-blog@1.3.14) (2020-03-06)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
 ## [1.3.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.3.12...gatsby-theme-blog@1.3.13) (2020-03-05)
 
 **Note:** Version bump only for package gatsby-theme-blog